requirements_rag.txt
streamlit
langchain
langchain_community
openai
python-dotenv
faiss-cpu
pypdf